// electron/types/ElectronApi.ts
export enum ElectronApi {
  GetAppPath = 'get-app-path',
  ExecCommand = 'exec-command',
  PathJoin = 'path-join',
  SpawnProcess = 'spawn-process',
  FileExists = 'file-exists',
  WriteFile = 'write-file',
  ReadDirectory = 'read-directory',
  DirExists = 'dir-exists',
  ReadBinaryFile = 'read-binary-file',
  CreateDirectory = 'create-directory',
  ReadFileAsString = 'read-file-as-string',
  WriteBinaryFile = 'write-binary-file',
  ShowOpenDialog = 'show-open-dialog',
  SpawnStdout = 'spawn-stdout',
  SpawnStderr = 'spawn-stderr',
  SpawnExit = 'spawn-exit',
  WatchDirs = 'watch-dirs',
  UnwatchDirs = 'unwatch-dirs',
  SubscribeWatchDirs = 'subscribe-watch-dirs',
  UnsubscribeUnwatchDirs = 'unsubscribe-watch-dirs',
}